# log,time: improve performance for writing a line to a log, add Time.format_rfc3339_micro/0