--- a/vlib/log/log.v
+++ b/vlib/log/log.v
@@ -24,7 +24,8 @@ pub enum TimeFormat {
 	tf_custom_format // 'MMMM Do YY N kk:mm:ss A' output like: January 1st 22 AD 13:45:33 PM
 }
 
-pub fn (l &Log) get_time_string_len() int {
+// get_time_string_len returns the length of the time string for the given time format
+fn (l &Log) get_time_string_len() int {
 	match l.time_format {
 		.tf_ss_micro { return 26 }
 		.tf_default { return 16 }
@@ -135,16 +136,17 @@ pub fn (mut l Log) reopen() ! {
 
 // log_file writes log line `s` with `level` to the log file.
 fn (mut l Log) log_file(s string, level Level) {
-	timestamp := l.time_format(time.utc())
-	e := tag_to_file(level, l.short_tag)
+	// timestamp := l.time_format(time.utc())
+	// e := tag_to_file(level, l.short_tag)
 
 	unsafe {
-		space := ' '
-		l.ofile.write_ptr(timestamp.str, timestamp.len)
-		l.ofile.write_ptr(space.str, space.len)
-		l.ofile.write_ptr(e.str, e.len)
-		l.ofile.write_ptr(space.str, space.len)
+		// space := ' '
+		// l.ofile.write_ptr(timestamp.str, timestamp.len)
+		// l.ofile.write_ptr(space.str, space.len)
+		// l.ofile.write_ptr(e.str, e.len)
+		// l.ofile.write_ptr(space.str, space.len)
 		l.ofile.write_ptr(s.str, s.len)
+		l.ofile.write_ptr('\n'.str, 1)
 	}
 	if l.always_flush {
 		l.flush()