--- a/vlib/log/log.v
+++ b/vlib/log/log.v
@@ -143,9 +143,14 @@ fn (mut l Log) log_file(s string, level Level) {
 		space := ' '
 		l.ofile.write_ptr(timestamp.str, timestamp.len)
 		l.ofile.write_ptr(space.str, space.len)
+
+		l.ofile.write_ptr('['.str, 1)
 		l.ofile.write_ptr(e.str, e.len)
+		l.ofile.write_ptr(']'.str, 1)
+
 		l.ofile.write_ptr(space.str, space.len)
 		l.ofile.write_ptr(s.str, s.len)
+
 		l.ofile.write_ptr('\n'.str, 1)
 	}
 	if l.always_flush {