--- a/vlib/log/log.v
+++ b/vlib/log/log.v
@@ -114,24 +114,28 @@ pub fn (mut l Log) reopen() ! {
 	}
 }
 
+const space = ' '
+const lbracket = '['
+const rbracket = ']'
+const newline = '\n'
+
 // log_file writes log line `s` with `level` to the log file.
 fn (mut l Log) log_file(s string, level Level) {
 	timestamp := l.time_format(time.now())
 	e := tag_to_file(level, l.short_tag)
 
 	unsafe {
-		space := ' '
 		l.ofile.write_ptr(timestamp.str, timestamp.len)
 		l.ofile.write_ptr(space.str, space.len)
 
-		l.ofile.write_ptr('['.str, 1)
+		l.ofile.write_ptr(lbracket.str, 1)
 		l.ofile.write_ptr(e.str, e.len)
-		l.ofile.write_ptr(']'.str, 1)
+		l.ofile.write_ptr(rbracket.str, 1)
 
 		l.ofile.write_ptr(space.str, space.len)
 		l.ofile.write_ptr(s.str, s.len)
 
-		l.ofile.write_ptr('\n'.str, 1)
+		l.ofile.write_ptr(newline.str, 1)
 	}
 	if l.always_flush {
 		l.flush()