--- a/vlib/time/time_test.v
+++ b/vlib/time/time_test.v
@@ -108,6 +108,18 @@ fn test_format_rfc3339() {
 	assert utc_res.contains('T')
 }
 
+fn test_format_rfc3339_micro() {
+	res := local_time_to_test.format_rfc3339_micro()
+	assert res.ends_with('23:42.123456Z')
+	assert res.starts_with('1980-07-1')
+	assert res.contains('T')
+
+	utc_res := utc_time_to_test.format_rfc3339_micro()
+	assert utc_res.ends_with('23:42.123456Z')
+	assert utc_res.starts_with('1980-07-1')
+	assert utc_res.contains('T')
+}
+
 fn test_format_rfc3339_nano() {
 	res := local_time_to_test.format_rfc3339_nano()
 	assert res.ends_with('23:42.123456789Z')