--- a/vlib/v/slow_tests/inout/dump_expression.out
+++ b/vlib/v/slow_tests/inout/dump_expression.out
@@ -10,7 +10,7 @@
             is_opened: false
         }
         output_target: console
-        time_format: tf_ss_micro
+        time_format: tf_rfc3339_micro
         custom_time_format: 'MMMM Do YY N kk:mm:ss A'
         short_tag: false
         always_flush: false