`v wipe-cache && v -prod crun log_.v > lala.log`

### Before

```sh
max_iterations: 1000000
SPENT  354.390 ms in log.info
SPENT  354.037 ms in log.error
SPENT   40.974 ms in println
```

### After

```sh
max_iterations: 1000000
SPENT  171.269 ms in log.info
SPENT  171.246 ms in log.error
SPENT   38.279 ms in println
```

```v
import log
import benchmark

const max_iterations := 1000000

fn main() {
	println('max_iterations: $max_iterations')

 	mut l := log.Log{}
	l.set_level(.info)
	l.set_full_logpath('./info.log')

	mut b := benchmark.start()

	for i := 0; i < max_iterations; i++ {
		l.info('Hello, World!')
	}

	b.measure('log.info')

	for i := 0; i < max_iterations; i++ {
		l.error('Hello, World!')
	}

	b.measure('log.error')

	for i := 0; i < max_iterations; i++ {
		println('Hello, World!')
	}
	b.measure('println')
}
```

*Note: if I remove time related format from log module

```sh
max_iterations: 1000000
SPENT   36.868 ms in log.info
SPENT   37.979 ms in log.error
SPENT   47.548 ms in println
```

*Note2: This was just a quick tweak I discovered out of curiosity. I'm not interested in making any major improvements right now, but there are several other tweaks and tests, such as using safe-threads when have logpath defined to log without interfering with the application's performance, and reusing time variables in some cases and managing cache.

<sub><a href="https://huly.app/guest/vlang-66f40c4d-a476b54c67-771fdd?token=eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJsaW5rSWQiOiI2NzFjZDY2M2YwMmQ4YTAzZmIzNDA1YTIiLCJndWVzdCI6InRydWUiLCJlbWFpbCI6IiNndWVzdEBoYy5lbmdpbmVlcmluZyIsIndvcmtzcGFjZSI6InctYWxleGFuZGVyLXZsYW5nLTY2ZjQwYzRkLWE0NzZiNTRjNjctNzcxZmRjIn0.8f3Z-Tt5fv0RmVMRr9tRrxFCHWny4QshjNx86T4GLcc">Huly&reg;: <b>V_0.6-21111</b></a></sub>