@spytheman 2024-10-26
> Write file in intervals should be good. like what happens here

This is good for a web server, but not as a general default.