@spytheman 2024-10-27
With the latest changes, on my machine:
![image](https://github.com/user-attachments/assets/3749cf5b-a588-4e7b-99e5-0aa4da7026bf)